pyproject.toml
[tool.poetry]
name = "image-stylizer"
version = "0.1.0"
description = ""
authors = ["Iratxe Moya <iratxe.moya@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.115.0"
uvicorn = "^0.31.0"
pillow = "^10.4.0"
torch = "^2.4.1"
diffusers = "^0.30.3"
transformers = "^4.45.1"
python-multipart = "^0.0.12"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
